requirements.txt
h5py
matplotlib
ninja
numpy
Pillow
tqdm
scipy
sklearn
six
seaborn
pandas
tensorboard
